requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.6
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
mysql-connector-python==8.0.26
mysqlclient
django-money==2.1.0
django-widget-tweaks==1.4.9
django-crispy-forms==1.13.0
mssql-django==1.0
pdfkit==1.00
django-phonenumber-field
python-decouple
django-filter